requirements.txt
pandas~=2.2.3
matplotlib~=3.7.3
yaml~=0.2.5
pyyaml~=6.0.1
setfit~=1.1.1
datasets~=3.3.2
nltk~=3.9.1
transformers~=4.49.0
events~=0.5
llama-index-core~=0.12.15
rapidfuzz~=3.12.2
numpy~=1.26.4
scikit-learn~=1.6.1
thefuzz~=0.22.1
plotly~=6.0.1
umap~=0.1.1
sentence-transformers~=3.4.1
requests~=2.32.3